db query change suggestion and retmax value
I tried to gather the entrezID of a list of ~2000 RefSeq entries. At the first, I stopped at the limit of the retrieved set size, 20, which is very low for the usual application of this kind of program, so, I set to the maximum value if a consistency check.
Then, I still had problems gathering many of the sequences which were not available on the gene database, so I had to change to nucleotide the query database.
An example is the MK966151 RefSeq. You can find its Entrez on the link:
https://eutils.ncbi.nlm.nih.gov/entrez/eutils/esearch.fcgi?db=nucleotide&term=MK966151
but you will not find it on:
https://eutils.ncbi.nlm.nih.gov/entrez/eutils/esearch.fcgi?db=gene&term=MK966151[ACCN]